feat: add sort to getProducts
