+++ types/index.ts
@@ -1,5 +1,5 @@
 export enum SortOption {
     CreatedAtDesc = 'newest',
-    PriceAsc = 'lower_price',
-    PriceDesc = 'higher_price'
+    PriceAsc = 'lower-price',
+    PriceDesc = 'higher-price'
 }
